[Update] script

+++ buildroot.sh
@@ -28,8 +28,8 @@ distributionDir=dist
 mkdir -p $distributionDir
 
 # コンフィグがあるならdistにコピーする
-configFileName=${1:-".config"}
-if [ -e $configFileName ]; then
+configFileName=${1:-""}
+if [ -n "$configFileName" -a -e "$configFileName" ]; then
     echo "Configuration file found. copy to ${distributionDir}..."
     cp $configFileName ${distributionDir}/.config
 fi
@@ -40,21 +40,40 @@ echo "While invoke make, please use instead:"
 echo "    make O=/${distributionDir}"
 docker run --rm -it -v ./${distributionDir}:/${distributionDir} $baseImageName
 
-# 終了後、コンフィグが変更されているか確認
-if [ -e ${distributionDir}/.config ]; then
-    cmp ${distributionDir}/.config $configFileName
-    if [ $? -ne 0 ]; then
-        echo "Copy generated config from /${distributionDir} to ./${distributionDir}."
-        echo "This operation needs super-user permission (password may be required)."
+# 構成ファイル名が渡されていなければ終了
+if [ -z "$configFileName" ]; then
+    echo "No configuration file passed."
+    echo "exit"
+    exit
+fi
 
-        # 変更されていれば、configファイルをdistディレクトリから元の場所に戻し、
-        # 所有者を現在のユーザとグループに書き換える
-        mv ${distributionDir}/.config $configFileName
-        if [ $? -ne 0 ]; then
-            sudo mv ${distributionDir}/.config $configFileName
-            currentUser=`whoami`
-            currentGroup=`id -g -n`
-            sudo chown $currentUser:$currentGroup $configFileName
-        fi
+# 構成ファイルを更新
+touch "$configFileName"
+
+# 構成ファイルが生成された?
+if [ ! -e "${distributionDir}/.config" ]; then
+    echo "No configuration file generated."
+    echo "exit"
+    exit
+fi
+
+# 渡された構成ファイルと現在手元にあるものを比較
+cmp "${distributionDir}/.config" "$configFileName" > /dev/null;isNotModified=$?
+if [ $isNotModified -ne 0 ];then
+    # 変更されていれば、configファイルをdistディレクトリから元の場所にコピー
+    echo "Copy generated config from /${distributionDir} to ./${distributionDir}."
+    echo "This operation needs super-user permission (password may be required)."
+    mv "${distributionDir}/.config" "$configFileName"
+    if [ $? -ne 0 ]; then
+        sudo mv "${distributionDir}/.config" "$configFileName"
+    fi
+
+    # 所有者を現在のユーザとグループに書き換える
+    # (Dockerコンテナが操作したファイルのownerがroot:rootに変わってしまう場合があるため)
+    currentUser=`whoami`
+    currentGroup=`id -g -n`
+    chown "$currentUser:$currentGroup" "$configFileName"
+    if [ $? -ne 0 ]; then
+        sudo chown "$currentUser:$currentGroup" "$configFileName"
     fi
 fi
